Began moving generic code out.
