+++ plugins/modules/acme_account.py
@@ -159,11 +159,12 @@ import base64
 from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.community.crypto.plugins.module_utils.acme import (
-    ModuleFailException,
     ACMEAccount,
     handle_standard_module_arguments,
     get_default_argspec,
 )
+
+from ansible_collections.community.crypto.plugins.module_utils.acme.errors import ModuleFailException
 
 
 def main():
