+++ plugins/modules/acme_certificate_revoke.py
@@ -120,12 +120,16 @@ RETURN = '''#'''
 from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.community.crypto.plugins.module_utils.acme import (
-    ModuleFailException,
     ACMEAccount,
-    nopad_b64,
-    pem_to_der,
     handle_standard_module_arguments,
     get_default_argspec,
+)
+
+from ansible_collections.community.crypto.plugins.module_utils.acme.errors import ModuleFailException
+
+from ansible_collections.community.crypto.plugins.module_utils.acme.utils import (
+    nopad_b64,
+    pem_to_der,
 )
