+++ plugins/modules/acme_inspect.py
@@ -246,11 +246,12 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils._text import to_native, to_bytes, to_text
 
 from ansible_collections.community.crypto.plugins.module_utils.acme import (
-    ModuleFailException,
     ACMEAccount,
     handle_standard_module_arguments,
     get_default_argspec,
 )
+
+from ansible_collections.community.crypto.plugins.module_utils.acme.errors import ModuleFailException
 
 
 def main():
